+++ R/drop_empty_cols.R
@@ -43,7 +43,7 @@ drop_empty_cols <- function(DT, copy = FALSE) {
               "Either ensure `DT` is a data.table or assert `copy = TRUE`.")
     }
     
-    isEmpty <- vapply(DT, function(x) all(is.na(x)), logical(1), USE.NAMES = FALSE)
+    isEmpty <- vapply(DT, function(x) anyNA(x) && all(is.na(x)), logical(1), USE.NAMES = FALSE)
     if (any(isEmpty)) {
       non_empty_cols <- which(!isEmpty)
       out <- DT[, non_empty_cols, drop = FALSE]